+++ src/Krystal/Stdlib/CHANGELOG.md
@@ -7,7 +7,7 @@ CHANGELOG
  * `VirtualEntity` since now throws `\RuntimeException` if a method being called is not starting with "get" or "set"
  * Since now `Dumper::dump()` renders objects if they implement `__toString()`
  * `VirtualEntity` since now can be treated as array and its properties can be accessed via snake case
- * Added `ArrayCollection` that simplifies work with static collections
+ * Added `ArrayCollection` and `ArrayGroupCollection` that simplifies work with static collections
  * Added `arrayDropdown()` in `Stdlib\ArrayUtils` to drop result-sets into partitions and prepare them for dropdowns
  * Added `arrayPartition()` in `Stdlib\ArrayUtils` to drop result-sets into partitions
  * Added `arrayRecovery()` in `Stdlib\ArrayUtils` to recover missing keys